requirements.txt
PyMySQL>=0.7.1
Flask>=0.11.1
flask_bootstrap>=3.3.7.0
Flask-Mail>=0.9.1
Flask-Principal>=0.4.0
Flask-Migrate>=2.0.0
Flask-Moment>=0.5.1
Flask-SQLAlchemy>=2.1
Flask-Script>=0.6.6
Flask-WTF>=0.13.1
Flask-Avatar>=0.1.2
Jinja2>=2.7.1
Mako>=0.9.1
MarkupSafe>=0.18
SQLAlchemy>=0.8.4
WTForms>=1.0.5
Werkzeug>=0.9.4
alembic>=0.6.2
blinker>=1.3
itsdangerous>=0.23
Flask-PageDown>=0.2.2
markdown>=2.6.7
bleach>=1.5.0
wheezy.captcha>=0.1.44
Flask-HTTPAuth>=3.2.1
ForgeryPy>=0.1
httpie>=0.9.6
coverage>=4.2
selenium>=3.0.2
nose>=1.3.7
Flask-WhooshAlchemyPlus>=0.7.5